--- .storybook/preview.ts
@@ -1,23 +1,0 @@
-import type { Preview } from '@storybook/nextjs-vite';
-import '@/app/globals.css';
-import './preview.css';
-
-const preview: Preview = {
-  parameters: {
-    controls: {
-      matchers: {
-        color: /(background|color)$/i,
-        date: /Date$/i,
-      },
-    },
-
-    a11y: {
-      // 'todo' - show a11y violations in the test UI only
-      // 'error' - fail CI on a11y violations
-      // 'off' - skip a11y checks entirely
-      test: 'todo',
-    },
-  },
-};
-
-export default preview;
